chore: Update npm dependencies and configure OpenAI API key
This commit updates the npm dependencies by adding the "axios" package with version "^1.7.0". Additionally, it configures the OpenAI API key in the nuxt.config.ts file using the runtimeConfig object. The apiKey is set to the value of the environment variable NUXT_APP_OPENAI_API_KEY or the default value "your-openai-api-key". These changes improve the functionality and security of the application.