chore: Add @nuxtjs/tailwindcss and shadcn-nuxt modules, configure shadcn options
This commit adds the "@nuxtjs/tailwindcss" and "shadcn-nuxt" modules to the project's dependencies in the package.json file. It also configures the "shadcn" options in the nuxt.config.ts file, setting the prefix to an empty string and the componentDir to "./components/ui". These changes enhance the project's styling capabilities and allow for easier management of UI components.